Separate Rate constructors implementations
