+++ rclcpp/include/rclcpp/rate.hpp
@@ -127,23 +127,11 @@ public:
 
   explicit Rate(
     const double rate,
-    Clock::SharedPtr clock = std::make_shared<Clock>(RCL_SYSTEM_TIME))
-  : clock_(clock), period_(0, 0), last_interval_(clock_->now())
-  {
-    if (rate <= 0.0) {
-      throw std::invalid_argument{"rate must be greater than 0"};
-    }
-    period_ = Duration::from_seconds(1.0 / rate);
-  }
+    Clock::SharedPtr clock = std::make_shared<Clock>(RCL_SYSTEM_TIME));
+
   explicit Rate(
     const Duration & period,
-    Clock::SharedPtr clock = std::make_shared<Clock>(RCL_SYSTEM_TIME))
-  : clock_(clock), period_(period), last_interval_(clock_->now())
-  {
-    if (period <= Duration(0, 0)) {
-      throw std::invalid_argument{"period must be greater than 0"};
-    }
-  }
+    Clock::SharedPtr clock = std::make_shared<Clock>(RCL_SYSTEM_TIME));
 
   virtual bool
   sleep();
@@ -172,25 +160,15 @@ private:
 class WallRate : public Rate
 {
 public:
-  explicit WallRate(const double rate)
-  : Rate(rate, std::make_shared<Clock>(RCL_STEADY_TIME))
-  {}
-
-  explicit WallRate(const Duration & period)
-  : Rate(period, std::make_shared<Clock>(RCL_STEADY_TIME))
-  {}
+  explicit WallRate(const double rate);
+  explicit WallRate(const Duration & period);
 };
 
 class ROSRate : public Rate
 {
 public:
-  explicit ROSRate(const double rate)
-  : Rate(rate, std::make_shared<Clock>(RCL_ROS_TIME))
-  {}
-
-  explicit ROSRate(const Duration & period)
-  : Rate(period, std::make_shared<Clock>(RCL_ROS_TIME))
-  {}
+  explicit ROSRate(const double rate);
+  explicit ROSRate(const Duration & period);
 };
 
 }  // namespace rclcpp
